Fix uppercase transformations of combo latin chars
Some cyrillic chars needs to be transformed into a combination of
latin chars. There are edge cases here for the case where a word
is composed only of uppercase or lowercase characters. The algorithm
performs a simple check to see if the next char in the word is also
an uppercase char and then proceeds to transform the combination of
latin chars to uppercase as well.